combined scroller and frame. Renamed ScrollContext to Scroller.Context. Moved deprecated components into separate directory until we're ready to kill them.
