+++ src/front/api/vendor.js
@@ -1,9 +1,11 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import * as classnames from 'classnames'
+import * as tunnels from 'react-tunnels'
 
 export default {
 	React,
 	ReactDOM,
 	classnames,
+	tunnels,
 }
