+++ src/front/ui/components/_deprecated/frame/style.scss
