+++ src/front/ui/components/_deprecated/split-view/index.js
