+++ src/front/ui/components/_deprecated/split-view/style.scss
