+++ src/front/ui/components/contexts/index.js
@@ -6,9 +6,6 @@ UIContext.displayName = 'UIContext'
 export const AppContext = createContext()
 AppContext.displayName = 'AppContext'
 
-export const StackContext = createContext()
-StackContext.displayName = 'StackContext'
-
 export const PageViewContext = createContext()
 PageViewContext.displayName = 'PageViewContext'
