--- src/front/ui/components/panel-parts/scroller/style.scss
@@ -1,12 +1,0 @@
-
-.fl-asst-scroller {
-    flex: 1 1 auto;
-    display: flex;
-    flex-direction: column;
-    overflow-y: scroll; /* has to be scroll, not auto */
-    -webkit-overflow-scrolling: touch;
-    overscroll-behavior: contain;
-    overflow: -moz-scrollbars-none;
-    -ms-overflow-style: none;
-    position: relative;
-}
